convert `useInserterMediaCategories` to selector and make private
